backend/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
drf_yasg==1.20.0
Faker==8.10.3
PyJWT==2.1.0
